feat(samples/saaanchez): remove certificate
This certificate, despite being weird and being
signed by a production key of AD, is in fact
genuine and used as part of the EU DGC Quality Assurance tests.

This is very unfortunate as these certificates
should have never been signed with a productive key
to begin with.

For this reason, this certificate isn't forged and
thus it is now removed with this commit.

References:
- https://github.com/denysvitali/covid-cert-analysis/pull/9
- https://github.com/ministero-salute/it-dgc-verificaC19-android/issues/185
- https://github.com/admin-ch/CovidCertificate-App-Android/issues/306
- https://github.com/eu-digital-green-certificates/dcc-quality-assurance/pull/183