feat(samples/pass-valido): remove certificate
Similarly to the previous commit, pass-valido is
now removed. The reason for that is that the
certificate was never a forged one to begin with but
was rather considered valid because the base
certificate (saaaanchez) is valid, and the
additional data at the end is simply ignored by
most of the verifier apps.